backend/requirements.txt
asgiref==3.5.2
black==22.8.0
boto3==1.22.8
botocore==1.25.8
click==8.1.3
dj-database-url==0.5.0
Django==4.1
django-cleanup==6.0.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
flake8==5.0.4
gunicorn==20.1.0
isort==5.10.1
jmespath==1.0.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
